Only patch openURL in test suite, not at runtime.
To avoid conflicts when using pydov together with OWSLib (f.ex. to call other services).
