+++ pydov/types/fields.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 """Module grouping all classes related to pydov field definitions."""
