+++ pydov/util/caching.py
@@ -7,7 +7,7 @@ import re
 import shutil
 import tempfile
 
-from pydov.util.dovutil import get_dov_xml
+from pydov.util.dovutil import build_dov_url, get_dov_xml
 from pydov.util.hooks import HookRunner
 
 
@@ -119,8 +119,8 @@ class AbstractFileCache(AbstractCache):
         self.max_age = max_age
 
         self._re_type_key = re.compile(
-            r'https?://(www|oefen|ontwikkel)\.dov\.vlaanderen\.be/'
-            r'data/([^ /]+)/([^.]+)')
+            build_dov_url('data/') + r'([^ /]+)/([^.]+)'
+        )
 
         try:
             if not os.path.exists(self.cachedir):
@@ -165,8 +165,8 @@ class AbstractFileCache(AbstractCache):
 
         """
         datatype = self._re_type_key.search(url)
-        if datatype and len(datatype.groups()) > 2:
-            return datatype.group(2), datatype.group(3)
+        if datatype and len(datatype.groups()) > 1:
+            return datatype.group(1), datatype.group(2)
 
     def _get_type_key_from_path(self, path):
         """Parse a filepath and return the datatype and object key.
